# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EnquiryInputAudience = typeof EnquiryInputAudience[keyof typeof EnquiryInputAudience];


export const EnquiryInputAudience = {
  Candidate: 'Candidate',
  Employer: 'Employer',
} as const;

export interface EnquiryInput {
  audience: EnquiryInputAudience;
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 5
     * @maxLength 40
     */
  phone: string;
  /**
     * @maxLength 160
     * @nullable
     */
  companyName?: string | null;
  /**
     * @minLength 2
     * @maxLength 180
     */
  areaOfInterest: string;
  /**
     * @minLength 10
     * @maxLength 5000
     */
  message: string;
}

export interface ApplicationInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 5
     * @maxLength 40
     */
  phone: string;
  /**
     * @minLength 2
     * @maxLength 160
     */
  currentLocation: string;
  /**
     * @minLength 2
     * @maxLength 160
     */
  role: string;
  /** @maxLength 80 */
  yearsOfExperience: string;
  /** @maxLength 300 */
  linkedin?: string;
  /** @maxLength 5000 */
  message?: string;
  resume: string;
}

export interface SubmissionResponse {
  success: boolean;
  message: string;
}