# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Website enquiry and application submissions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /enquiries:
    post:
      operationId: sendEnquiry
      tags: [contact]
      summary: Send a website enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryInput"
      responses:
        "200":
          description: Enquiry accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          description: Invalid enquiry
        "429":
          description: Too many requests
        "500":
          description: Mail delivery failed
  /applications:
    post:
      operationId: submitApplication
      tags: [contact]
      summary: Submit a career application with a resume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "200":
          description: Application accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          description: Invalid application
        "413":
          description: Resume too large
        "429":
          description: Too many requests
        "500":
          description: Mail delivery failed
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    EnquiryInput:
      type: object
      required:
        - audience
        - fullName
        - email
        - phone
        - areaOfInterest
        - message
      properties:
        audience:
          type: string
          enum: [Candidate, Employer]
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          maxLength: 254
        phone:
          type: string
          minLength: 5
          maxLength: 40
        companyName:
          type: ["string", "null"]
          maxLength: 160
        areaOfInterest:
          type: string
          minLength: 2
          maxLength: 180
        message:
          type: string
          minLength: 10
          maxLength: 5000
    ApplicationInput:
      type: object
      required:
        - fullName
        - email
        - phone
        - currentLocation
        - role
        - yearsOfExperience
        - resume
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          maxLength: 254
        phone:
          type: string
          minLength: 5
          maxLength: 40
        currentLocation:
          type: string
          minLength: 2
          maxLength: 160
        role:
          type: string
          minLength: 2
          maxLength: 160
        yearsOfExperience:
          type: string
          maxLength: 80
        linkedin:
          type: string
          maxLength: 300
        message:
          type: string
          maxLength: 5000
        resume:
          type: string
    SubmissionResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string