# lib/api-zod/src/generated/types/enquiryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EnquiryInputAudience } from './enquiryInputAudience';

export interface EnquiryInput {
  audience: EnquiryInputAudience;
  /**
     * @minLength 2
     * @maxLength 120
     */
  fullName: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 5
     * @maxLength 40
     */
  phone: string;
  /**
     * @maxLength 160
     * @nullable
     */
  companyName?: string | null;
  /**
     * @minLength 2
     * @maxLength 180
     */
  areaOfInterest: string;
  /**
     * @minLength 10
     * @maxLength 5000
     */
  message: string;
}